Changed Interaction system: strategy pattern applied
