+++ Assets/Mechanics/Spells/Fireboll.asset
@@ -14,4 +14,9 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   spellName: 
   description: 
+  area: 0
+  projectileSpeed: 0
+  healthCost: 0
+  staminaCost: 0
+  manaCost: 0
   spellTicks: []
