+++ Assets/Scripts/Interaction/ActionAfterInteraction.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: b706050bd1e74e059aa952fe20fa07a2
+timeCreated: 1681796925
