+++ Assets/Scripts/Interaction/ActionAfterInteraction/ChestActionInteraction.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: efb3d2f0996b4c94a9038244730c0fd8
+timeCreated: 1681797147
