+++ Assets/Scripts/Interaction/ActionAfterInteraction/DoorActionInteraction.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: f625d92483e342b9acce8d95d62b48a3
+timeCreated: 1681796961
