+++ Assets/Scripts/Interaction/IActionAfterInterection.cs
@@ -1,0 +1,7 @@
+﻿
+using UnityEngine;
+
+public abstract class IActionAfterInterection : MonoBehaviour
+{
+    public abstract void Action(GameObject gameObject);
+}
