+++ Assets/Scripts/Interaction/IActionAfterInterection.cs.meta
@@ -1,0 +1,3 @@
+﻿fileFormatVersion: 2
+guid: d1e9553b889b4d539f911402c225b489
+timeCreated: 1681796840
