+++ Assets/Scripts/Interaction/InteractionAction.cs
@@ -22,24 +22,11 @@ public class InteractionAction : MonoBehaviour
 
             case "Door":
 
-                Transform oldTransform = gameObject.GetComponent<Transform>();
-                Transform parentTransform = oldTransform.parent;
-                ReplacementObject replacementObject = gameObject.GetComponent<ReplacementObject>();
-                GameObject newGameObject = replacementObject.NewGameObject;
-                Vector3 oldPosition = oldTransform.localPosition;
-                var resultPosition = parentTransform.position + oldPosition + replacementObject.CorrectionVector;
-                Quaternion oldQuaternion = oldTransform.rotation;
-
-                Destroy(gameObject);
-                var createdObject = Instantiate(original: newGameObject, position: resultPosition,
-                    rotation: oldQuaternion, parent: parentTransform);
-                break;
+                
             
             case "Chest":
 
-                SpriteRenderer renderer = gameObject.GetComponent<SpriteRenderer>();
-                SpriteChanger newSprite = gameObject.GetComponent<SpriteChanger>();
-                renderer.sprite = newSprite.SpriteForChange;
+                
                 
                 break;
         }
