+++ Assets/Scripts/Interaction/InteractionObject.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class InteractionObject : MonoBehaviour
 {
     [SerializeField] private float timeCost;
-    [SerializeField] private string actionName;
+    [SerializeField] private IActionAfterInterection actionInteraction;
     [SerializeField] private new GameObject gameObject;
     public float TimeCost => timeCost;
 
@@ -15,8 +15,6 @@ public class InteractionObject : MonoBehaviour
 
     public void InteractionCompleted()
     {
-        InteractionAction.Action(actionName, gameObject);
+        actionInteraction.Action(gameObject);
     }
-    
-    
 }
