--- Logs/shadercompiler-UnityShaderCompiler.exe0.log
@@ -1,3 +1,0 @@
-Base path: 'D:/Unity/2022.2.14f1/Editor/Data', plugins path 'D:/Unity/2022.2.14f1/Editor/Data/PlaybackEngines'
-Cmd: initializeCompiler
-
